pyproject.toml
[project]
name = "customer-support-backend"
version = "0.1.0"
description = ""
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "fastapi==0.115.12",
  "bcrypt==3.2.0",
  "email-validator==2.2.0",
  "jose==1.0.0",
  "passlib==1.7.4",
  "psycopg2-binary==2.9.10",
  "pydantic==2.11.3",
  "pydantic-settings==2.9.1",
  "pydantic-core==2.33.1",
  "python-dotenv==1.1.0",
  "python-jose==3.4.0",
  "requests==2.32.3",
  "SQLAlchemy==2.0.40",
  "sse-starlette==2.2.1",
  "starlette==0.46.2",
  "urllib3==2.4.0",
  "uvicorn==0.34.2"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
